Re-add update_client logic
