+++ relayer-spike/src/foreign_client.rs
@@ -3,8 +3,6 @@ use crate::chain::{Chain, SignedHeader, ConsensusState, MembershipProof};
 
 #[derive(Debug)]
 pub enum ForeignClientError {
-    VerificationError(),
-    HeaderMismatch(),
 }
 
 pub struct ForeignClientConfig {
@@ -39,52 +37,4 @@ impl ForeignClient {
         })
     }
 
-    /* TODO: This has to move to the Chain
-    pub fn update(&mut self, src_target_height: Height) -> Result<Height, ForeignClientError> {
-        let (mut src_consensus_state, mut dst_membership_proof) =
-            self.dst_chain.consensus_state(self.src_chain.chain_id, src_target_height);
-
-        let dst_sh = self.dst_chain.get_header(dst_membership_proof.height + 1);
-        // type verifyMembership = (root: CommitmentRoot, proof: CommitmentProof, path: CommitmentPath, value: Value) => boolean (ICS-023)
-        if ! verify_consensus_state_inclusion(&src_consensus_state, &dst_membership_proof, &(dst_sh.header.app_hash)) {
-            // Error: Destination chain provided invalid consensus_state
-            return Err(ForeignClientError::VerificationError())
-        }
-
-        // verify client_state on self
-        if self.src_chain.get_header(src_consensus_state.height).header == src_consensus_state.signed_header.header {
-            return Err(ForeignClientError::HeaderMismatch())
-        }
-
-        while src_consensus_state.height < src_target_height {
-            let src_signed_headers = self.src_chain.get_minimal_set(src_consensus_state.height, src_target_height);
-
-            // This might fail semantically due to competing relayers
-            // Even if this fails, we need to continue
-            self.dst_chain.submit(vec![create_client_update_datagram(src_signed_headers)]);
-
-            let (src_consensus_state, dst_membership_proof) = self.dst_chain.consensus_state(self.src_chain.chain_id, src_target_height);
-            let dst_sh = self.dst_chain.get_header(dst_membership_proof.height + 1);
-            if ! verify_consensus_state_inclusion(&src_consensus_state, &dst_membership_proof, &(dst_sh.header.app_hash)) {
-                // Error: Destination chain provided invalid client_state
-                return Err(ForeignClientError::VerificationError())
-            }
-
-            if self.src_chain.get_header(src_consensus_state.height).header == src_consensus_state.signed_header.header {
-                // Error: consesus_state isn't verified by self light client
-                return  Err(ForeignClientError::HeaderMismatch())
-            }
-        }
-
-        return Ok(src_target_height)
-    }
-    */
-}
-
-fn verify_consensus_state_inclusion(_consensus_state: &ConsensusState, _membership_proof: &MembershipProof, _hash: &Hash) -> bool {
-    return true
-}
-
-fn create_client_update_datagram(_header: Vec<SignedHeader>) -> Datagram  {
-    return Datagram::NoOp()
 }
